Added purgecss functionality to the gulpfile
Moved "Gulp Packages" to top as this needs to run before lazypipes can be used. A lazypipe is used in the paths section now for user clarity in adding their own purgecss related paths.  I've been unsuccesul in calling a path directly into the 'content' part of purgecss using any other method so far, so adding the whole process as a lazypipe under the current paths seemed the best alternative for now.

There is a bug with gulp watch and lazypipe which causes active watching to fail if a lazypipe is the last pipe used in a series of pipes. To get around this, I've added ".pipe(gutil.noop());" after the relevant lazypipes. This does nothing but by simply being there, this is a workaround for the bug.

As purgecss can read comments in a CSS file that indicate whitelisted CSS that should not be removed under any circumstances, 'sourceComments' in the sass pipe has been set to false. We then delay the minifying of CSS until after purgecss has run IF purgecss is set to run so that it doesn't needlessly run twice. As such, CSS minification was added to a new lazypipe to be called when needed.

Purging of CSS needs to read the outputted CSS, not the original sass file, to work correctly. Therefore it is set to run as a non-asynchronous task after all other tasks have run.